Support DCN layout
DCN has multiple controplane networks, making it different from any
other layout we had until now.

A compute is usually connected to "a public network", and to a specific,
"local" controlplane, while another compute would be connected to a
different controlplane.

In addition to this, with Linux bridge name length limit, we must also
check in the "translation map" in case a network in libvirt isn't named
like the "interesting" ones used in the networking_mapper output.

This patch makes use of that translation when we generate the DHCP
entries - that was apparently the missing part, since we generate the
DNS entries already based on the networking_mapper output, with a
"match" for the `ctlplane` pattern, allowing to capture any network
starting with that specific pattern.